requirements.txt
pandas
sqlalchemy
psycopg2-binary
dbt-postgres
python-dotenv